diyabet/diyabet_takip/requirements.txt
PyQt5==5.15.9
mysql-connector-python==8.0.33
cryptography==41.0.3
pandas==2.0.3
matplotlib==3.7.2
python-dateutil==2.8.2
bcrypt==4.0.1
python-dotenv==1.0.0 